Set warnings as errors for the generator, the samples, and the tests.

+++ samples/06_InitDepthBuffer/06_InitDepthBuffer.cpp
@@ -85,7 +85,7 @@ int main( int /*argc*/, char ** /*argv*/ )
       }
       typeBits >>= 1;
     }
-    assert( typeIndex != ~0 );
+    assert( typeIndex != uint32_t( ~0 ) );
     vk::UniqueDeviceMemory depthMemory =
       device->allocateMemoryUnique( vk::MemoryAllocateInfo( memoryRequirements.size, typeIndex ) );
